utilisation de longs pour stocker kes montant plutot que de double, ca evite les erreurs d'arrondi...

+++ dao/src/main/java/com/excilys/formation/bank/dao/impl/CompteDAOHibernateImpl.java
@@ -50,7 +50,7 @@ public class CompteDAOHibernateImpl implements CompteDAO {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public final void updateSolde(String compteId, double difference) {
+	public final void updateSolde(String compteId, long difference) {
 		Compte compte = loadCompteById(compteId);
 		compte.setSolde(compte.getSolde() + difference);
 		sessionFactory.getCurrentSession().update(compte);
