+++ dao/src/test/java/com/excilys/formation/bank/dao/CompteDaoTest.java
@@ -65,8 +65,7 @@ public class CompteDaoTest extends
 	 */
 	@Test
 	public final void updateCompteEnleverArgent() {
-		Double soldeInitial = compteDAO.loadCompteById("compte1user1")
-				.getSolde();
+		long soldeInitial = compteDAO.loadCompteById("compte1user1").getSolde();
 		compteDAO.updateSolde("compte1user1", -20);
 		assertThat(compteDAO.loadCompteById("compte1user1").getSolde())
 				.isEqualTo(soldeInitial - 20);
@@ -77,8 +76,7 @@ public class CompteDaoTest extends
 	 */
 	@Test
 	public final void updateCompteAjouterArgent() {
-		Double soldeInitial = compteDAO.loadCompteById("compte1user1")
-				.getSolde();
+		long soldeInitial = compteDAO.loadCompteById("compte1user1").getSolde();
 		compteDAO.updateSolde("compte1user1", 20);
 		assertThat(compteDAO.loadCompteById("compte1user1").getSolde())
 				.isEqualTo(soldeInitial + 20);
